pyproject.toml
[project]
name = "blogregator"
version = "0.1.0"
description = "Blogregator is a tool for automatically scraping blogs, extracting metadata from posts, and updating you when a new post drops."
readme = "README.md"
authors = [
    { name = "Aman Choudhri", email = "amanchoudhri@gmail.com" }
]
requires-python = ">=3.11"
dependencies = [
    "apscheduler>=3.10.4",
    "beautifulsoup4>=4.13.4",
    "fastapi>=0.109.0",
    "litellm>=1.69.2",
    "playwright>=1.57.0",
    "psycopg2-binary>=2.9.11",
    "python-json-logger>=2.0.7",
    "requests>=2.32.3",
    "ruff>=0.14.10",
    "tenacity>=8.2.3",
    "typer>=0.15.3",
    "uvicorn[standard]>=0.27.0",
]

[project.optional-dependencies]
dev = ["ruff>=0.8.0"]

[project.scripts]
blogregator = "blogregator.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort (import sorting)
    "N",   # pep8-naming
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
]
ignore = [
    "E501",  # line too long (handled by formatter)
]

[tool.ruff.lint.isort]
known-first-party = ["blogregator"]
